mem: Removed circular include ref
If BasicLink.hh is modified, the style checker forces a reordering of 
the includes, which results in build errors because it ends up including 
Topology.hh before including its xxxParams.hh files, which include 
forward declarations of the BasicLink family of classes, and so 
Topology.hh throws errors that BasicLink etc. are not declared. 

Change-Id: I664a0652e53f0cc61763c2190a980c655b85d397
Signed-off-by: Ryan Gambord <gambordr@oregonstate.edu>
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/17270
Reviewed-by: Nikos Nikoleris <nikos.nikoleris@arm.com>
Maintainer: Nikos Nikoleris <nikos.nikoleris@arm.com>

Former-commit-id: 9e22a2ab603d743b187108986cfbeba07c0c0b8d